update use of 7-bit mode
Moved 7-bit mode defines to td64.h because they are used outside of the 7-bit mode.

When fewer than minimum values to use 7-bit mode of 16, don't accumulate high bit when reasonable. Main loop keeps this in because time required is minimal.

When fewer than 24 input values, but greater than or equal to minimum values of 16 to use 7-bit mode, use 6% as minimum compression for compression modes used prior to 7-bit mode.